site survey for madwifi
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@3978 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/services/site_survey_broadcom.c
@@ -45,7 +45,7 @@ struct site_survey_list
   unsigned char BSSID[18];
   uint8 channel;		/* Channel no. */
   int16 RSSI;			/* receive signal strength (in dBm) */
-  int8 phy_noise;		/* noise (in dBm) */
+  int16 phy_noise;		/* noise (in dBm) */
   uint16 beacon_period;		/* units are Kusec */
   uint16 capability;		/* Capability information */
   uint rate_count;		/* # rates in this set */
